requirements.txt
streamlit
cohere
faiss-cpu
pdfplumber
numpy

